requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
autopep8==2.0.4
jsbeautifier==1.14.11
beautifulsoup4==4.12.3
